backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
qdrant-client==1.8.0
google-generativeai==0.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-slugify==8.0.1
aiofiles==23.2.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
markdown==3.5.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2